Restructure files. Add scope in Property
