+++ src/components/windows/ListObject.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useForceUpdate } from '../../hooks/useForceUpdate';
 import { EveryObjectProperty } from '../../lib/properties/EveryObjectProperty';
-import { EditorObject } from '../Editor';
+import { EditorObject } from "../EditorObject";
 
 export function ListObject({ property }: { property: EveryObjectProperty; }): JSX.Element {
     return <div>
