+++ src/lib/properties/EveryObjectProperty.ts
@@ -1,10 +1,11 @@
 import { fabric } from 'fabric';
-import { BaseState, EditorObject, FabricContext } from '../../components/Editor';
+import { FabricContext } from "../../components/FabricContext";
+import { EditorObject } from "../../components/EditorObject";
 import { Property } from '../core/Property';
 
 
 export class EveryObjectProperty extends Property<EditorObject[]> {
-    onInit(canvas: fabric.Canvas, context: FabricContext<BaseState>) {
+    onInit(canvas: fabric.Canvas, context: FabricContext) {
         canvas.on('object:added', () => this.change$.next(this.getValue()));
         canvas.on('object:removed', () => this.change$.next(this.getValue()));
     }
