+++ src/lib/utilities/getRandomUid.tsx
@@ -1,0 +1,4 @@
+
+export function getRandomUid(): string {
+    return Math.random().toString(36).substring(2, 15) + Math.random().toString(36).substring(2, 15);
+}
